Split out commands and refactor/cleanup
