+++ src/listeners/command-listener.ts
@@ -1,11 +1,17 @@
 import { REST } from "@discordjs/rest";
 import { Routes } from "discord-api-types/v9";
-import { AutocompleteInteraction, CacheType, Interaction } from "discord.js";
+import {
+  AutocompleteInteraction,
+  CacheType,
+  CommandInteraction,
+  Interaction,
+} from "discord.js";
 import { auctionCommand } from "../features/spell-auctions/command";
 import { clientId, guildId, token } from "../config";
-import { bankHourCommand } from "../features/bank-hours/command";
+import { setBankHourCommand } from "../features/bank-hours/set-command";
+import { removeBankHour } from "../features/bank-hours/remove-command";
 
-const commands = [auctionCommand, bankHourCommand];
+const commands = [auctionCommand, setBankHourCommand, removeBankHour];
 
 export const registerCommands = () => {
   const rest = new REST({ version: "9" }).setToken(token);
@@ -23,17 +29,22 @@ export const commandListener = async (interaction: Interaction<CacheType>) => {
     return;
   }
 
-  if (!interaction.isCommand()) return;
-
-  const { commandName } = interaction;
-  commands
-    .find((c) => c.name === commandName)
-    ?.listen(interaction)
-    .then((success) =>
-      console.log(`Received /${commandName}: ${success ? "success" : "failed"}`)
-    )
-    .catch(console.error);
+  if (interaction.isCommand()) {
+    getCommand(interaction)
+      ?.listen(interaction)
+      .then((success) =>
+        console.log(
+          `Received /${interaction.commandName}: ${
+            success ? "success" : "failed"
+          }`
+        )
+      )
+      .catch(console.error);
+  }
 };
 
-const getCommand = (interaction: AutocompleteInteraction<CacheType>) =>
-  commands.find((c) => c.name === interaction.commandName);
+const getCommand = (
+  interaction:
+    | CommandInteraction<CacheType>
+    | AutocompleteInteraction<CacheType>
+) => commands.find((c) => c.name === interaction.commandName);
